Split integration and unit tests
