+++ tests/unit/test_dakara_server.py
