+++ tests/unit/test_diff_generator.py
