+++ tests/unit/test_directory_lister.py
