+++ tests/unit/test_similarity_calculator.py
